little tweaks and fixes

+++ src/main/java/skylands/api/BalanceHandler.java
@@ -2,20 +2,25 @@ package skylands.api;
 
 import net.minecraft.entity.player.PlayerEntity;
 
+/**
+ * Represents player balance handler
+ * probably can be used for compatibility
+ * with other economy mods
+ */
 public interface BalanceHandler {
 
-	int getBalance(PlayerEntity player);
+	double getBalance(PlayerEntity player);
 
-	void setBalance(PlayerEntity player, int balance);
+	void setBalance(PlayerEntity player, double balance);
 
-	default void increaseBalance(PlayerEntity player, int increment) {
-		int balance = this.getBalance(player);
+	default void increaseBalance(PlayerEntity player, double increment) {
+		double balance = this.getBalance(player);
 		balance = balance + increment;
 		this.setBalance(player, balance);
 	}
 
-	default void decreaseBalance(PlayerEntity player, int decrement) {
-		int balance = this.getBalance(player);
+	default void decreaseBalance(PlayerEntity player, double decrement) {
+		double balance = this.getBalance(player);
 		balance = balance - decrement;
 		this.setBalance(player, balance);
 	}
